OAK-6081 - Indexing tooling via oak-run
Specify an explicit subdirectory for storing the output results

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1796979 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-run/src/main/java/org/apache/jackrabbit/oak/index/IndexCommand.java
@@ -159,7 +159,9 @@ public class IndexCommand implements Command {
     }
 
     private static void setupDirectories(IndexOptions indexOpts) throws IOException {
-        FileUtils.cleanDirectory(indexOpts.getOutDir());
+        if (indexOpts.getOutDir().exists()) {
+            FileUtils.cleanDirectory(indexOpts.getOutDir());
+        }
         cleanWorkDir(indexOpts.getWorkDir());
     }
